fiz fuzz zimage

+++ fuzz/fuzzers/fuzz_zimage.rs
@@ -10,6 +10,6 @@ use std::io::Cursor;
 
 fuzz_target!(|data: &[u8]| { // fuzzed code goes here
     let mut kimage = Cursor::new(data);
-    let mem = GuestMemory::new(&[(GuestAddress(0), data.len() + 0x1000)]).unwrap();
+    let mem = GuestMemory::new(&[(GuestAddress(0), data.len() as u64 + 0x1000)]).unwrap();
     let _ = kernel_loader::load_kernel(&mem, GuestAddress(0), &mut kimage);
 });
